Move createServiceTester into its own module
