+++ services/npm/npm-type-definitions.tester.js
@@ -1,9 +1,9 @@
 'use strict'
 
 const Joi = require('joi')
-const ServiceTester = require('../service-tester')
+const createServiceTester = require('../create-service-tester')
 
-const t = ServiceTester.forThisService()
+const t = createServiceTester()
 module.exports = t
 
 const isTypeDefinition = Joi.string().regex(
