+++ services/npm/npm-version.tester.js
@@ -1,10 +1,10 @@
 'use strict'
 
 const Joi = require('joi')
-const ServiceTester = require('../service-tester')
+const createServiceTester = require('../create-service-tester')
 const { isSemver } = require('../test-validators')
 
-const t = ServiceTester.forThisService()
+const t = createServiceTester()
 module.exports = t
 
 t.create('gets the package version of left-pad')
