+++ services/uptimerobot/uptimerobot-status.tester.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const Joi = require('joi')
-const ServiceTester = require('../service-tester')
+const createServiceTester = require('../create-service-tester')
 
 const isUptimeStatus = Joi.string().valid(
   'paused',
@@ -12,7 +12,7 @@ const isUptimeStatus = Joi.string().valid(
 )
 const { invalidJSON } = require('../response-fixtures')
 
-const t = ServiceTester.forThisService()
+const t = createServiceTester()
 module.exports = t
 
 t.create('Uptime Robot: Status (valid)')
